Use "double" instead of "long" data type for scores
This allows us to support fractional numbers for scores.
The precision of double should be enough for our scoring use cases.

+++ graylog2-server/src/main/java/org/graylog/events/event/Event.java
@@ -24,7 +24,7 @@ import org.joda.time.DateTime;
 
 import java.util.List;
 import java.util.Map;
-import java.util.OptionalLong;
+import java.util.OptionalDouble;
 
 public interface Event extends Indexable {
     @Override
@@ -102,9 +102,9 @@ public interface Event extends Indexable {
 
     EventReplayInfo getReplayInfo();
 
-    OptionalLong getScore(String name);
+    OptionalDouble getScore(String name);
 
-    void setScore(String name, long riskScore);
+    void setScore(String name, double riskScore);
 
     EventDto toDto();
